links adjustments, components/files/page structure(hooks added), styles adjustments(responsive designing)
