+++ frontend/src/components/ArtistProfile/styles/UploadPortfolio.css
@@ -118,6 +118,7 @@ button:nth-child(1) {
   display: flex;
   justify-content: space-evenly;
   align-items: center;
+  position: relative;
 }
 
 .add-photos {
@@ -128,4 +129,6 @@ button:nth-child(1) {
   cursor: pointer;
   font-size: 30px;
   margin-bottom: 20px;
+  position: absolute;
+  right: 0;
 }
