+++ frontend/src/components/Home/styles/hero.css
@@ -57,7 +57,6 @@
       padding: 4px 0;
       font-size: 1.5rem;
       font-family: Verdana, Geneva, Tahoma, sans-serif;
-      margin-left: 7rem;
 
       &&:hover {
         background-color: #b31b1b;
@@ -117,6 +116,10 @@
       p {
         font-size: clamp(2rem, 3vw, 3rem);
         padding-left: 5rem;
+      }
+
+      .hero-button {
+        margin-left: 7rem;
       }
     }
   }
