+++ frontend/src/components/NavBar.jsx
@@ -1,61 +1,31 @@
-import React, { useEffect, useState } from "react";
+import React, { useState } from "react";
 import "./NavBar.css";
-import { NavLink, useLocation, useNavigate } from "react-router-dom";
+import { NavLink } from "react-router-dom";
 import { FiMenu, FiX } from "react-icons/fi";
 import { useContext } from "react";
 import { UserContext } from "../context/ContextProvider";
-import { HiChat, HiLogout } from "react-icons/hi";
-import useProfileImageUpload from "../hooks/useProfileImageUpload";
 import { FaCamera } from "react-icons/fa";
 import DropdownNav from "./Navigation/DropdownNav";
-import CustomSlideImages from "./Navigation/CustomSlideImages";
+import ProfileImageUpload from "./uploading/ProfileImageUpload";
+import useLocationFinder from "../hooks/useLocationFinder";
 
 function NavBar() {
-  const location = useLocation();
-  const { user, logout } = useContext(UserContext);
+  useLocationFinder();
+  const { user } = useContext(UserContext);
   const [activeDropdown, setActiveDropdown] = useState(null);
   const [isUploadDialogOpen, setUploadDialogOpen] = useState(false);
-  const navigate = useNavigate();
-
-  const {
-    file,
-    preview,
-    loading,
-    error,
-    response,
-    handleSelectFile,
-    handleUpload,
-  } = useProfileImageUpload();
 
   const toggleDropdown = (dropdownName) => {
     setActiveDropdown(activeDropdown === dropdownName ? null : dropdownName);
   };
 
+  //image
   const toggleUploadDialog = () => {
     setUploadDialogOpen(!isUploadDialogOpen);
   };
 
-  const handleProfileImageUpload = async () => {
-    if (file) {
-      await handleUpload(); // Upload the image using the custom hook
-      toggleUploadDialog();
-    }
-  };
-
-  useEffect(() => {
-    if (location.hash) {
-      const element = document.querySelector(location.hash);
-      if (element) {
-        element.scrollIntoView({ behavior: "smooth" });
-      }
-    } else if (location.pathname === "/") {
-      window.scrollTo({ top: 0, behavior: "smooth" });
-    }
-  }, [location]);
-
-  const logoutHandler = async () => {
-    await logout();
-    navigate("/login");
+  const handleProfileImageSuccess = async () => {
+    toggleUploadDialog();
   };
 
   return (
@@ -97,24 +67,7 @@ function NavBar() {
                 className="camera-icon"
               />
             </div>
-
-            {activeDropdown === "profile" && (
-              <div className="nav-profile-path open-profile-dropdown">
-                <div className="account-dropdown-menu">
-                  <CustomSlideImages />
-                  <h2 className="customer-name">
-                    {`Welcome, ${user.firstName} ${user.lastName}!`}
-                  </h2>
-                  <DropdownNav />
-                  <NavLink className="chat-link">
-                    <HiChat /> &nbsp; Messages
-                  </NavLink>
-                  <button className="nav-button-logout" onClick={logoutHandler}>
-                    <HiLogout size={21} /> &nbsp; Logout
-                  </button>
-                </div>
-              </div>
-            )}
+            {activeDropdown === "profile" && <DropdownNav />}
           </div>
         ) : (
           <div className="button-container">
@@ -126,42 +79,10 @@ function NavBar() {
 
         {/* Dialog for uploading the profile image */}
         {isUploadDialogOpen && (
-          <div className="upload-dialog">
-            <div className="upload-dialog-content">
-              <h2>Upload Profile Image</h2>
-              <div className="preview-profile">
-                {" "}
-                {preview && <img src={preview} style={{ width: "50px" }} />}
-              </div>
-              <input
-                type="file"
-                accept="image/*"
-                onChange={handleSelectFile}
-                disabled={loading}
-                className="file-input"
-              />
-              {loading && <p>Uploading...</p>}
-              {error && <p style={{ color: "red" }}>{error}</p>}
-              {response && (
-                <p style={{ color: "green" }}>Image uploaded successfully!</p>
-              )}
-              <div className="profileImage-upload-buttons">
-                <button
-                  onClick={handleProfileImageUpload}
-                  disabled={loading}
-                  className="uploadimage-button"
-                >
-                  Upload Image
-                </button>
-                <button
-                  onClick={toggleUploadDialog}
-                  className="uploadimage-button"
-                >
-                  Cancel
-                </button>
-              </div>
-            </div>
-          </div>
+          <ProfileImageUpload
+            onClose={toggleUploadDialog}
+            onSuccess={handleProfileImageSuccess}
+          />
         )}
       </div>
     </>
